Switch to App::Cmd
...because we're basically an App::Cmd scriptie already.  See how much code we
get to delete right off the bat, and not have to maintain anymore? =)

Note that this is not the complete conversion, just the removal of files we
can delete outright and the introduction of new ones, as well as the
conversion of the script.  The actual command classes are moved and refactored
in successive commits.