Add .travis.yaml, the better to test us with
Right now it's only configured to use a smattering of perls (the ones I think
we're going to hit, and the latest stable) and the default install.  The
default install is sufficient to start, but we *may* need to extend it at some
point to handle testing a full cycle.